requirements.txt
aerospike==6.0.0
appnope==0.1.2
argon2-cffi==20.1.0
async-generator==1.10
attrs==21.2.0
autobahn==19.11.2
Automat==20.2.0
backcall==0.2.0
backtrader==1.9.76.123
bleach==3.3.0
cachetools==4.2.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coloredlogs==15.0.1
constantly==15.1.0
cryptography==3.4.7
cycler==0.10.0
dateparser==1.0.0
decorator==5.0.9
defusedxml==0.7.1
dill==0.3.4
entrypoints==0.3
enum34==1.1.10
et-xmlfile==1.1.0
ffn==0.3.6
future==0.18.2
google-auth==2.0.1
google-auth-oauthlib==0.4.5
gspread==4.0.1
humanfriendly==10.0
hyperlink==21.0.0
idna==2.10
imageio==2.9.0
incremental==21.3.0
install==1.3.5
ipdb==0.13.9
ipykernel==5.5.5
ipython==7.24.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kaleido==0.2.1
kiteconnect==3.9.4
kiwisolver==1.3.1
llvmlite==0.36.0
lml==0.1.0
lxml==4.6.3
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
multitasking==0.0.9
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.0
numba==0.53.1
numpy==1.20.3
oauthlib==3.1.1
openpyxl==3.0.7
packaging==20.9
panda==0.3.1
pandas==1.2.4
pandas-datareader==0.9.0
pandocfilters==1.4.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
plotly==5.1.0
prometheus-client==0.10.1
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyexcel==0.6.6
pyexcel-io==0.6.4
Pygments==2.9.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPrind==2.11.3
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.1.0
qtconsole==5.1.0
QtPy==1.9.0
regex==2021.4.4
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
schedule==1.1.0
scikit-learn==0.24.2
scipy==1.7.0
Send2Trash==1.5.0
service-identity==21.1.0
six==1.16.0
TA-Lib==0.4.24
tabulate==0.8.9
tenacity==7.0.0
terminado==0.10.0
testpath==0.5.0
texttable==1.6.4
threadpoolctl==2.1.0
toml==0.10.2
tornado==6.1
tqdm==4.61.1
traitlets==5.0.5
Twisted==21.2.0
txaio==21.2.1
tzlocal==2.1
urllib3==1.26.5
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xlrd==2.0.1
yfinance==0.1.59
zope.interface==5.4.0
